Update Mammoth.dmm
absolutely FUCKING SCRAMBLED THE MAP. Rearranged it so it's got a bit more space to jam buildings into. There's a ton of new, empty buildings to fill. I await Ridder's Hotel rework and Sen's Casino to put across from it. But for now, take this as a small hint of what i'm trying to do, as it's subject to change